chore(menu): move stories into examples
